Add datetime for resource attribute type
This change adds a new type of attribute for resource: datetime.
