+++ releasenotes/notes/datetime-resource-attribute-type-1e627a686568f72a.yaml
@@ -1,0 +1,4 @@
+---
+features:
+  - |
+    A new date type is available for resource type attribute: datetime.
