+++ setup.cfg
@@ -93,6 +93,7 @@ gnocchi.indexer.sqlalchemy.resource_type_attribute =
     uuid = gnocchi.indexer.sqlalchemy_extension:UUIDSchema
     number = gnocchi.indexer.sqlalchemy_extension:NumberSchema
     bool = gnocchi.indexer.sqlalchemy_extension:BoolSchema
+    datetime = gnocchi.indexer.sqlalchemy_extension:DatetimeSchema
 
 gnocchi.storage =
     swift = gnocchi.storage.swift:SwiftStorage
